telegram_bot_gpt-main/requirements.txt
python-telegram-bot>=20.4
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.1
httpx>=0.24.0
